some changes in api

+++ server/controllers/product.js
@@ -1,9 +1,39 @@
-const express = require('express');
-const router = express.Router();
-const Product = require('./models/Product');
-router.post('/add-product', async (req, res) => {
+import Product from "../models/Product.js";
+   
+  // @desc   add product
+  // route   POST api product/addProduct
+  // access  Private
+
+  export async function addProduct(req, res) {
     try {
         const { productName, description, image, category, sellingPrice, Mrp, quantity } = req.body;
+
+        if(!productName || !category || !sellingPrice || !Mrp || !quantity){
+            return res.status(404).json({ 
+                success: false,
+                message: 'fill all detail of Product' 
+            });
+        }
+
+        if(!description){
+            description = "";
+        }
+
+        const productPic = req?.files?.productPicture;
+        if(!productPic){
+            return res.status(404).json({
+                success: false,
+                message: "file not found in body",
+            });
+        }
+
+        const img = await cloudinaryFileUpload(
+            productPic,
+            process.env.FOLDER_NAME,
+            1000,
+            100,
+        );
+
         const product = new Product({
             productName,
             description,
@@ -14,15 +44,37 @@ router.post('/add-product', async (req, res) => {
             quantity
         });
         const savedProduct = await product.save();
-        res.json({ message: 'Product added successfully', product: savedProduct });
+        return res.status(200).json({
+            success: true,
+            product: savedProduct,
+            message: "Product added successfully",
+        });
     } catch (error) {
-        res.status(500).json({ message: 'Failed to add product', error: error.message });
+        return res.status(500).json({ 
+            success: false,
+            message: 'Failed to add product',
+            error: error.message 
+        });
     }
-});
-router.put('/edit-product/:id', async (req, res) => {
+}
+
+  // @desc   edit product
+  // route   PUT api product/editProduct
+  // access  Private
+
+  export async function editProduct (req, res) {
     try {
         const productId = req.params.id;
+
         const { productName, description, image, category, sellingPrice, Mrp, quantity } = req.body;
+
+        if(!productId){
+            return res.status(404).json({ 
+                success: false,
+                message: 'ProductId not found' 
+            });
+        }
+
         const updatedProduct = await Product.findByIdAndUpdate(productId, {
             productName,
             description,
@@ -31,25 +83,59 @@ router.put('/edit-product/:id', async (req, res) => {
             sellingPrice,
             Mrp,
             quantity
-        }, { new: true });
+        },{ new: true });
+
         if (!updatedProduct) {
-            return res.status(404).json({ message: 'Product not found' });
+            return res.status(404).json({ 
+                success: false,
+                message: 'Product not found' 
+            });
         }
-        res.json({ message: 'Product edited successfully', product: updatedProduct });
+        return res.status(200).json({ 
+            success: true,
+            message: 'Product edited successfully', 
+            product: updatedProduct 
+        });
     } catch (error) {
-        res.status(500).json({ message: 'Failed to edit product', error: error.message });
+        return res.status(500).json({ 
+            success : false,
+            message: 'Failed to edit product', 
+            error: error.message 
+        });
     }
-});
-router.delete('/delete-product/:id', async (req, res) => {
+}
+
+  // @desc   delete product
+  // route   DELETE api product/deleteProduct
+  // access  Private
+
+  export async function deleteProduct(req, res) {
     try {
         const productId = req.params.id;
+
+        if(!productId){
+            return res.status(404).json({ 
+                success: false,
+                message: 'ProductId not found' 
+            });
+        }
+
         const deletedProduct = await Product.findByIdAndDelete(productId);
         if (!deletedProduct) {
-            return res.status(404).json({ message: 'Product not found' });
+            return res.status(404).json({ 
+                success: false,
+                message: 'Product not found' 
+            });
         }
-        res.json({ message: 'Product deleted successfully' });
+        return res.status(200).json({ 
+            success: true,
+            message: 'Product deleted successfully' 
+        });
     } catch (error) {
-        res.status(500).json({ message: 'Failed to delete product', error: error.message });
+        return res.status(500).json({ 
+            success: false,
+            message: 'Failed to delete product',
+            error: error.message 
+        });
     }
-});
-module.exports = router;
+}
